aws-perpetual-range/aws-perpetual-range-app/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
requests-aws4auth==1.1.2
boto3==latest